requirements.txt
# Phase 2 dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
python-dateutil>=2.8.2

# Phase 3 additions
streamlit==1.28.0
openai==1.6.1 
faiss-cpu==1.7.4
sentence-transformers==2.2.2
crewai==0.28.0
langchain>=0.1.10,<0.2.0
python-dotenv==1.0.0
numpy==1.24.0
plotly==5.18.0